+++ src/main/kotlin/click/metroeye/api/config/security/SecurityConfig.kt
@@ -1,8 +1,8 @@
-package click.metroeye.api.config
+package click.metroeye.api.config.security
 
 import click.metroeye.api.constants.ErrorCode
-import click.metroeye.api.infrastructure.security.filter.BearerAuthenticationWebFilter
-import click.metroeye.api.infrastructure.security.manager.BearerAuthenticationManager
+import click.metroeye.api.config.security.BearerAuthenticationWebFilter
+import click.metroeye.api.config.security.BearerAuthenticationManager
 import click.metroeye.api.presentation.v1.dto.response.ApiResponse
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.context.annotation.Bean
@@ -25,7 +25,8 @@ class SecurityConfig {
     @Order(1)
     fun swaggerSecurityFilterChain(http: ServerHttpSecurity): SecurityWebFilterChain {
         return http
-            .securityMatcher(ServerWebExchangeMatchers.pathMatchers(
+            .securityMatcher(
+                ServerWebExchangeMatchers.pathMatchers(
                 "/",
                 "/swagger-ui.html",
                 "/swagger-ui/**",
